+++ src/lib.rs
@@ -1,4 +1,7 @@
 #![feature(field_init_shorthand, static_in_const, type_ascription)]
+#![cfg_attr(test, feature(plugin))]
+#![cfg_attr(test, plugin(clippy))]
+
 extern crate multistr;
 
 mod data_line;