+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 This project follows semantic versioning.
 
+# 0.2.0
+
+* [changed] `DataParseError` now includes the full text of failed hosts
+* [added] `DataParseError` and `LineReadError` now implement `Error`
+* [added] `HostsFile` now has a `minimal_lines` option
+* [fixed] `BadHost` now works properly
+
 # 0.1.0
 
 * Initial release.